Struct aws_sdk_lightsail::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for Lightsail_20161128.

This client allows ergonomic access to a Lightsail_20161128-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn allocate_static_ip(&self) -> AllocateStaticIp<C, M, R>

pub fn attach_certificate_to_distribution(
&self
) -> AttachCertificateToDistribution<C, M, R>

pub fn attach_disk(&self) -> AttachDisk<C, M, R>

pub fn attach_instances_to_load_balancer(
&self
) -> AttachInstancesToLoadBalancer<C, M, R>

pub fn attach_load_balancer_tls_certificate(
&self
) -> AttachLoadBalancerTlsCertificate<C, M, R>

pub fn attach_static_ip(&self) -> AttachStaticIp<C, M, R>

pub fn close_instance_public_ports(&self) -> CloseInstancePublicPorts<C, M, R>

pub fn copy_snapshot(&self) -> CopySnapshot<C, M, R>

pub fn create_bucket(&self) -> CreateBucket<C, M, R>

pub fn create_bucket_access_key(&self) -> CreateBucketAccessKey<C, M, R>

pub fn create_certificate(&self) -> CreateCertificate<C, M, R>

pub fn create_cloud_formation_stack(&self) -> CreateCloudFormationStack<C, M, R>

pub fn create_contact_method(&self) -> CreateContactMethod<C, M, R>

pub fn create_container_service(&self) -> CreateContainerService<C, M, R>

pub fn create_container_service_deployment(
&self
) -> CreateContainerServiceDeployment<C, M, R>

pub fn create_container_service_registry_login(
&self
) -> CreateContainerServiceRegistryLogin<C, M, R>

pub fn create_disk(&self) -> CreateDisk<C, M, R>

pub fn create_disk_from_snapshot(&self) -> CreateDiskFromSnapshot<C, M, R>

pub fn create_disk_snapshot(&self) -> CreateDiskSnapshot<C, M, R>

pub fn create_distribution(&self) -> CreateDistribution<C, M, R>

pub fn create_domain(&self) -> CreateDomain<C, M, R>

pub fn create_domain_entry(&self) -> CreateDomainEntry<C, M, R>

pub fn create_instances(&self) -> CreateInstances<C, M, R>

pub fn create_instances_from_snapshot(
&self
) -> CreateInstancesFromSnapshot<C, M, R>

pub fn create_instance_snapshot(&self) -> CreateInstanceSnapshot<C, M, R>

pub fn create_key_pair(&self) -> CreateKeyPair<C, M, R>

pub fn create_load_balancer(&self) -> CreateLoadBalancer<C, M, R>

pub fn create_load_balancer_tls_certificate(
&self
) -> CreateLoadBalancerTlsCertificate<C, M, R>

pub fn create_relational_database(&self) -> CreateRelationalDatabase<C, M, R>

pub fn create_relational_database_from_snapshot(
&self
) -> CreateRelationalDatabaseFromSnapshot<C, M, R>

pub fn create_relational_database_snapshot(
&self
) -> CreateRelationalDatabaseSnapshot<C, M, R>

pub fn delete_alarm(&self) -> DeleteAlarm<C, M, R>

pub fn delete_auto_snapshot(&self) -> DeleteAutoSnapshot<C, M, R>

pub fn delete_bucket(&self) -> DeleteBucket<C, M, R>

pub fn delete_bucket_access_key(&self) -> DeleteBucketAccessKey<C, M, R>

pub fn delete_certificate(&self) -> DeleteCertificate<C, M, R>

pub fn delete_contact_method(&self) -> DeleteContactMethod<C, M, R>

pub fn delete_container_image(&self) -> DeleteContainerImage<C, M, R>

pub fn delete_container_service(&self) -> DeleteContainerService<C, M, R>

pub fn delete_disk(&self) -> DeleteDisk<C, M, R>

pub fn delete_disk_snapshot(&self) -> DeleteDiskSnapshot<C, M, R>

pub fn delete_distribution(&self) -> DeleteDistribution<C, M, R>

pub fn delete_domain(&self) -> DeleteDomain<C, M, R>

pub fn delete_domain_entry(&self) -> DeleteDomainEntry<C, M, R>

pub fn delete_instance(&self) -> DeleteInstance<C, M, R>

pub fn delete_instance_snapshot(&self) -> DeleteInstanceSnapshot<C, M, R>

pub fn delete_key_pair(&self) -> DeleteKeyPair<C, M, R>

pub fn delete_known_host_keys(&self) -> DeleteKnownHostKeys<C, M, R>

pub fn delete_load_balancer(&self) -> DeleteLoadBalancer<C, M, R>

pub fn delete_load_balancer_tls_certificate(
&self
) -> DeleteLoadBalancerTlsCertificate<C, M, R>

pub fn delete_relational_database(&self) -> DeleteRelationalDatabase<C, M, R>

pub fn delete_relational_database_snapshot(
&self
) -> DeleteRelationalDatabaseSnapshot<C, M, R>

pub fn detach_certificate_from_distribution(
&self
) -> DetachCertificateFromDistribution<C, M, R>

pub fn detach_disk(&self) -> DetachDisk<C, M, R>

pub fn detach_instances_from_load_balancer(
&self
) -> DetachInstancesFromLoadBalancer<C, M, R>

pub fn detach_static_ip(&self) -> DetachStaticIp<C, M, R>

pub fn disable_add_on(&self) -> DisableAddOn<C, M, R>

pub fn download_default_key_pair(&self) -> DownloadDefaultKeyPair<C, M, R>

pub fn enable_add_on(&self) -> EnableAddOn<C, M, R>

pub fn export_snapshot(&self) -> ExportSnapshot<C, M, R>

pub fn get_active_names(&self) -> GetActiveNames<C, M, R>

pub fn get_alarms(&self) -> GetAlarms<C, M, R>

pub fn get_auto_snapshots(&self) -> GetAutoSnapshots<C, M, R>

pub fn get_blueprints(&self) -> GetBlueprints<C, M, R>

pub fn get_bucket_access_keys(&self) -> GetBucketAccessKeys<C, M, R>

pub fn get_bucket_bundles(&self) -> GetBucketBundles<C, M, R>

pub fn get_bucket_metric_data(&self) -> GetBucketMetricData<C, M, R>

pub fn get_buckets(&self) -> GetBuckets<C, M, R>

pub fn get_bundles(&self) -> GetBundles<C, M, R>

pub fn get_certificates(&self) -> GetCertificates<C, M, R>

pub fn get_cloud_formation_stack_records(
&self
) -> GetCloudFormationStackRecords<C, M, R>

pub fn get_contact_methods(&self) -> GetContactMethods<C, M, R>

pub fn get_container_api_metadata(&self) -> GetContainerAPIMetadata<C, M, R>

pub fn get_container_images(&self) -> GetContainerImages<C, M, R>

pub fn get_container_log(&self) -> GetContainerLog<C, M, R>

pub fn get_container_service_deployments(
&self
) -> GetContainerServiceDeployments<C, M, R>

pub fn get_container_service_metric_data(
&self
) -> GetContainerServiceMetricData<C, M, R>

pub fn get_container_service_powers(&self) -> GetContainerServicePowers<C, M, R>

pub fn get_container_services(&self) -> GetContainerServices<C, M, R>

pub fn get_disk(&self) -> GetDisk<C, M, R>

pub fn get_disks(&self) -> GetDisks<C, M, R>

pub fn get_disk_snapshot(&self) -> GetDiskSnapshot<C, M, R>

pub fn get_disk_snapshots(&self) -> GetDiskSnapshots<C, M, R>

pub fn get_distribution_bundles(&self) -> GetDistributionBundles<C, M, R>

pub fn get_distribution_latest_cache_reset(
&self
) -> GetDistributionLatestCacheReset<C, M, R>

pub fn get_distribution_metric_data(&self) -> GetDistributionMetricData<C, M, R>

pub fn get_distributions(&self) -> GetDistributions<C, M, R>

pub fn get_domain(&self) -> GetDomain<C, M, R>

pub fn get_domains(&self) -> GetDomains<C, M, R>

pub fn get_export_snapshot_records(&self) -> GetExportSnapshotRecords<C, M, R>

pub fn get_instance(&self) -> GetInstance<C, M, R>

pub fn get_instance_access_details(&self) -> GetInstanceAccessDetails<C, M, R>

pub fn get_instance_metric_data(&self) -> GetInstanceMetricData<C, M, R>

pub fn get_instance_port_states(&self) -> GetInstancePortStates<C, M, R>

pub fn get_instances(&self) -> GetInstances<C, M, R>

pub fn get_instance_snapshot(&self) -> GetInstanceSnapshot<C, M, R>

pub fn get_instance_snapshots(&self) -> GetInstanceSnapshots<C, M, R>

pub fn get_instance_state(&self) -> GetInstanceState<C, M, R>

pub fn get_key_pair(&self) -> GetKeyPair<C, M, R>

pub fn get_key_pairs(&self) -> GetKeyPairs<C, M, R>

pub fn get_load_balancer(&self) -> GetLoadBalancer<C, M, R>

pub fn get_load_balancer_metric_data(
&self
) -> GetLoadBalancerMetricData<C, M, R>

pub fn get_load_balancers(&self) -> GetLoadBalancers<C, M, R>

pub fn get_load_balancer_tls_certificates(
&self
) -> GetLoadBalancerTlsCertificates<C, M, R>

pub fn get_operation(&self) -> GetOperation<C, M, R>

pub fn get_operations(&self) -> GetOperations<C, M, R>

pub fn get_operations_for_resource(&self) -> GetOperationsForResource<C, M, R>

pub fn get_regions(&self) -> GetRegions<C, M, R>

pub fn get_relational_database(&self) -> GetRelationalDatabase<C, M, R>

pub fn get_relational_database_blueprints(
&self
) -> GetRelationalDatabaseBlueprints<C, M, R>

pub fn get_relational_database_bundles(
&self
) -> GetRelationalDatabaseBundles<C, M, R>

pub fn get_relational_database_events(
&self
) -> GetRelationalDatabaseEvents<C, M, R>

pub fn get_relational_database_log_events(
&self
) -> GetRelationalDatabaseLogEvents<C, M, R>

pub fn get_relational_database_log_streams(
&self
) -> GetRelationalDatabaseLogStreams<C, M, R>

pub fn get_relational_database_master_user_password(
&self
) -> GetRelationalDatabaseMasterUserPassword<C, M, R>

pub fn get_relational_database_metric_data(
&self
) -> GetRelationalDatabaseMetricData<C, M, R>

pub fn get_relational_database_parameters(
&self
) -> GetRelationalDatabaseParameters<C, M, R>

pub fn get_relational_databases(&self) -> GetRelationalDatabases<C, M, R>

pub fn get_relational_database_snapshot(
&self
) -> GetRelationalDatabaseSnapshot<C, M, R>

pub fn get_relational_database_snapshots(
&self
) -> GetRelationalDatabaseSnapshots<C, M, R>

pub fn get_static_ip(&self) -> GetStaticIp<C, M, R>

pub fn get_static_ips(&self) -> GetStaticIps<C, M, R>

pub fn import_key_pair(&self) -> ImportKeyPair<C, M, R>

pub fn is_vpc_peered(&self) -> IsVpcPeered<C, M, R>

pub fn open_instance_public_ports(&self) -> OpenInstancePublicPorts<C, M, R>

pub fn peer_vpc(&self) -> PeerVpc<C, M, R>

pub fn put_alarm(&self) -> PutAlarm<C, M, R>

pub fn put_instance_public_ports(&self) -> PutInstancePublicPorts<C, M, R>

pub fn reboot_instance(&self) -> RebootInstance<C, M, R>

pub fn reboot_relational_database(&self) -> RebootRelationalDatabase<C, M, R>

pub fn register_container_image(&self) -> RegisterContainerImage<C, M, R>

pub fn release_static_ip(&self) -> ReleaseStaticIp<C, M, R>

pub fn reset_distribution_cache(&self) -> ResetDistributionCache<C, M, R>

pub fn send_contact_method_verification(
&self
) -> SendContactMethodVerification<C, M, R>

pub fn set_ip_address_type(&self) -> SetIpAddressType<C, M, R>

pub fn set_resource_access_for_bucket(
&self
) -> SetResourceAccessForBucket<C, M, R>

pub fn start_instance(&self) -> StartInstance<C, M, R>

pub fn start_relational_database(&self) -> StartRelationalDatabase<C, M, R>

pub fn stop_instance(&self) -> StopInstance<C, M, R>

pub fn stop_relational_database(&self) -> StopRelationalDatabase<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn test_alarm(&self) -> TestAlarm<C, M, R>

pub fn unpeer_vpc(&self) -> UnpeerVpc<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_bucket(&self) -> UpdateBucket<C, M, R>

pub fn update_bucket_bundle(&self) -> UpdateBucketBundle<C, M, R>

pub fn update_container_service(&self) -> UpdateContainerService<C, M, R>

pub fn update_distribution(&self) -> UpdateDistribution<C, M, R>

pub fn update_distribution_bundle(&self) -> UpdateDistributionBundle<C, M, R>

pub fn update_domain_entry(&self) -> UpdateDomainEntry<C, M, R>

pub fn update_load_balancer_attribute(
&self
) -> UpdateLoadBalancerAttribute<C, M, R>

pub fn update_relational_database(&self) -> UpdateRelationalDatabase<C, M, R>

pub fn update_relational_database_parameters(
&self
) -> UpdateRelationalDatabaseParameters<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self