Module aws_sdk_lightsail::error

Modules

access_denied_exception

See AccessDeniedException

account_setup_in_progress_exception

See AccountSetupInProgressException

invalid_input_exception

See InvalidInputException

not_found_exception

See NotFoundException

operation_failure_exception

See OperationFailureException

service_exception

See ServiceException

unauthenticated_exception

See UnauthenticatedException

Structs

AccessDeniedException

Lightsail throws this exception when the user cannot be authenticated or uses invalid

AccountSetupInProgressException

Lightsail throws this exception when an account is still in the setup in progress

AllocateStaticIpError

AttachCertificateToDistributionError

AttachDiskError

AttachInstancesToLoadBalancerError

AttachLoadBalancerTlsCertificateError

AttachStaticIpError

CloseInstancePublicPortsError

CopySnapshotError

CreateBucketAccessKeyError

CreateBucketError

CreateCertificateError

CreateCloudFormationStackError

CreateContactMethodError

CreateContainerServiceDeploymentError

CreateContainerServiceError

CreateContainerServiceRegistryLoginError

CreateDiskError

CreateDiskFromSnapshotError

CreateDiskSnapshotError

CreateDistributionError

CreateDomainEntryError

CreateDomainError

CreateInstanceSnapshotError

CreateInstancesError

CreateInstancesFromSnapshotError

CreateKeyPairError

CreateLoadBalancerError

CreateLoadBalancerTlsCertificateError

CreateRelationalDatabaseError

CreateRelationalDatabaseFromSnapshotError

CreateRelationalDatabaseSnapshotError

DeleteAlarmError

DeleteAutoSnapshotError

DeleteBucketAccessKeyError

DeleteBucketError

DeleteCertificateError

DeleteContactMethodError

DeleteContainerImageError

DeleteContainerServiceError

DeleteDiskError

DeleteDiskSnapshotError

DeleteDistributionError

DeleteDomainEntryError

DeleteDomainError

DeleteInstanceError

DeleteInstanceSnapshotError

DeleteKeyPairError

DeleteKnownHostKeysError

DeleteLoadBalancerError

DeleteLoadBalancerTlsCertificateError

DeleteRelationalDatabaseError

DeleteRelationalDatabaseSnapshotError

DetachCertificateFromDistributionError

DetachDiskError

DetachInstancesFromLoadBalancerError

DetachStaticIpError

DisableAddOnError

DownloadDefaultKeyPairError

EnableAddOnError

ExportSnapshotError

GetActiveNamesError

GetAlarmsError

GetAutoSnapshotsError

GetBlueprintsError

GetBucketAccessKeysError

GetBucketBundlesError

GetBucketMetricDataError

GetBucketsError

GetBundlesError

GetCertificatesError

GetCloudFormationStackRecordsError

GetContactMethodsError

GetContainerAPIMetadataError

GetContainerImagesError

GetContainerLogError

GetContainerServiceDeploymentsError

GetContainerServiceMetricDataError

GetContainerServicePowersError

GetContainerServicesError

GetDiskError

GetDiskSnapshotError

GetDiskSnapshotsError

GetDisksError

GetDistributionBundlesError

GetDistributionLatestCacheResetError

GetDistributionMetricDataError

GetDistributionsError

GetDomainError

GetDomainsError

GetExportSnapshotRecordsError

GetInstanceAccessDetailsError

GetInstanceError

GetInstanceMetricDataError

GetInstancePortStatesError

GetInstanceSnapshotError

GetInstanceSnapshotsError

GetInstanceStateError

GetInstancesError

GetKeyPairError

GetKeyPairsError

GetLoadBalancerError

GetLoadBalancerMetricDataError

GetLoadBalancerTlsCertificatesError

GetLoadBalancersError

GetOperationError

GetOperationsError

GetOperationsForResourceError

GetRegionsError

GetRelationalDatabaseBlueprintsError

GetRelationalDatabaseBundlesError

GetRelationalDatabaseError

GetRelationalDatabaseEventsError

GetRelationalDatabaseLogEventsError

GetRelationalDatabaseLogStreamsError

GetRelationalDatabaseMasterUserPasswordError

GetRelationalDatabaseMetricDataError

GetRelationalDatabaseParametersError

GetRelationalDatabaseSnapshotError

GetRelationalDatabaseSnapshotsError

GetRelationalDatabasesError

GetStaticIpError

GetStaticIpsError

ImportKeyPairError

InvalidInputException

Lightsail throws this exception when user input does not conform to the validation rules

IsVpcPeeredError

NotFoundException

Lightsail throws this exception when it cannot find a resource.

OpenInstancePublicPortsError

OperationFailureException

Lightsail throws this exception when an operation fails to execute.

PeerVpcError

PutAlarmError

PutInstancePublicPortsError

RebootInstanceError

RebootRelationalDatabaseError

RegisterContainerImageError

ReleaseStaticIpError

ResetDistributionCacheError

SendContactMethodVerificationError

ServiceException

A general service exception.

SetIpAddressTypeError

SetResourceAccessForBucketError

StartInstanceError

StartRelationalDatabaseError

StopInstanceError

StopRelationalDatabaseError

TagResourceError

TestAlarmError

UnauthenticatedException

Lightsail throws this exception when the user has not been authenticated.

UnpeerVpcError

UntagResourceError

UpdateBucketBundleError

UpdateBucketError

UpdateContainerServiceError

UpdateDistributionBundleError

UpdateDistributionError

UpdateDomainEntryError

UpdateLoadBalancerAttributeError

UpdateRelationalDatabaseError

UpdateRelationalDatabaseParametersError

Enums

AllocateStaticIpErrorKind

AttachCertificateToDistributionErrorKind

AttachDiskErrorKind

AttachInstancesToLoadBalancerErrorKind

AttachLoadBalancerTlsCertificateErrorKind

AttachStaticIpErrorKind

CloseInstancePublicPortsErrorKind

CopySnapshotErrorKind

CreateBucketAccessKeyErrorKind

CreateBucketErrorKind

CreateCertificateErrorKind

CreateCloudFormationStackErrorKind

CreateContactMethodErrorKind

CreateContainerServiceDeploymentErrorKind

CreateContainerServiceErrorKind

CreateContainerServiceRegistryLoginErrorKind

CreateDiskErrorKind

CreateDiskFromSnapshotErrorKind

CreateDiskSnapshotErrorKind

CreateDistributionErrorKind

CreateDomainEntryErrorKind

CreateDomainErrorKind

CreateInstanceSnapshotErrorKind

CreateInstancesErrorKind

CreateInstancesFromSnapshotErrorKind

CreateKeyPairErrorKind

CreateLoadBalancerErrorKind

CreateLoadBalancerTlsCertificateErrorKind

CreateRelationalDatabaseErrorKind

CreateRelationalDatabaseFromSnapshotErrorKind

CreateRelationalDatabaseSnapshotErrorKind

DeleteAlarmErrorKind

DeleteAutoSnapshotErrorKind

DeleteBucketAccessKeyErrorKind

DeleteBucketErrorKind

DeleteCertificateErrorKind

DeleteContactMethodErrorKind

DeleteContainerImageErrorKind

DeleteContainerServiceErrorKind

DeleteDiskErrorKind

DeleteDiskSnapshotErrorKind

DeleteDistributionErrorKind

DeleteDomainEntryErrorKind

DeleteDomainErrorKind

DeleteInstanceErrorKind

DeleteInstanceSnapshotErrorKind

DeleteKeyPairErrorKind

DeleteKnownHostKeysErrorKind

DeleteLoadBalancerErrorKind

DeleteLoadBalancerTlsCertificateErrorKind

DeleteRelationalDatabaseErrorKind

DeleteRelationalDatabaseSnapshotErrorKind

DetachCertificateFromDistributionErrorKind

DetachDiskErrorKind

DetachInstancesFromLoadBalancerErrorKind

DetachStaticIpErrorKind

DisableAddOnErrorKind

DownloadDefaultKeyPairErrorKind

EnableAddOnErrorKind

ExportSnapshotErrorKind

GetActiveNamesErrorKind

GetAlarmsErrorKind

GetAutoSnapshotsErrorKind

GetBlueprintsErrorKind

GetBucketAccessKeysErrorKind

GetBucketBundlesErrorKind

GetBucketMetricDataErrorKind

GetBucketsErrorKind

GetBundlesErrorKind

GetCertificatesErrorKind

GetCloudFormationStackRecordsErrorKind

GetContactMethodsErrorKind

GetContainerAPIMetadataErrorKind

GetContainerImagesErrorKind

GetContainerLogErrorKind

GetContainerServiceDeploymentsErrorKind

GetContainerServiceMetricDataErrorKind

GetContainerServicePowersErrorKind

GetContainerServicesErrorKind

GetDiskErrorKind

GetDiskSnapshotErrorKind

GetDiskSnapshotsErrorKind

GetDisksErrorKind

GetDistributionBundlesErrorKind

GetDistributionLatestCacheResetErrorKind

GetDistributionMetricDataErrorKind

GetDistributionsErrorKind

GetDomainErrorKind

GetDomainsErrorKind

GetExportSnapshotRecordsErrorKind

GetInstanceAccessDetailsErrorKind

GetInstanceErrorKind

GetInstanceMetricDataErrorKind

GetInstancePortStatesErrorKind

GetInstanceSnapshotErrorKind

GetInstanceSnapshotsErrorKind

GetInstanceStateErrorKind

GetInstancesErrorKind

GetKeyPairErrorKind

GetKeyPairsErrorKind

GetLoadBalancerErrorKind

GetLoadBalancerMetricDataErrorKind

GetLoadBalancerTlsCertificatesErrorKind

GetLoadBalancersErrorKind

GetOperationErrorKind

GetOperationsErrorKind

GetOperationsForResourceErrorKind

GetRegionsErrorKind

GetRelationalDatabaseBlueprintsErrorKind

GetRelationalDatabaseBundlesErrorKind

GetRelationalDatabaseErrorKind

GetRelationalDatabaseEventsErrorKind

GetRelationalDatabaseLogEventsErrorKind

GetRelationalDatabaseLogStreamsErrorKind

GetRelationalDatabaseMasterUserPasswordErrorKind

GetRelationalDatabaseMetricDataErrorKind

GetRelationalDatabaseParametersErrorKind

GetRelationalDatabaseSnapshotErrorKind

GetRelationalDatabaseSnapshotsErrorKind

GetRelationalDatabasesErrorKind

GetStaticIpErrorKind

GetStaticIpsErrorKind

ImportKeyPairErrorKind

IsVpcPeeredErrorKind

OpenInstancePublicPortsErrorKind

PeerVpcErrorKind

PutAlarmErrorKind

PutInstancePublicPortsErrorKind

RebootInstanceErrorKind

RebootRelationalDatabaseErrorKind

RegisterContainerImageErrorKind

ReleaseStaticIpErrorKind

ResetDistributionCacheErrorKind

SendContactMethodVerificationErrorKind

SetIpAddressTypeErrorKind

SetResourceAccessForBucketErrorKind

StartInstanceErrorKind

StartRelationalDatabaseErrorKind

StopInstanceErrorKind

StopRelationalDatabaseErrorKind

TagResourceErrorKind

TestAlarmErrorKind

UnpeerVpcErrorKind

UntagResourceErrorKind

UpdateBucketBundleErrorKind

UpdateBucketErrorKind

UpdateContainerServiceErrorKind

UpdateDistributionBundleErrorKind

UpdateDistributionErrorKind

UpdateDomainEntryErrorKind

UpdateLoadBalancerAttributeErrorKind

UpdateRelationalDatabaseErrorKind

UpdateRelationalDatabaseParametersErrorKind