Module aws_sdk_lightsail::model

Modules

access_key

See AccessKey

access_key_last_used

See AccessKeyLastUsed

access_rules

See AccessRules

add_on

See AddOn

add_on_request

See AddOnRequest

alarm

See Alarm

attached_disk

See AttachedDisk

auto_snapshot_add_on_request

See AutoSnapshotAddOnRequest

auto_snapshot_details

See AutoSnapshotDetails

availability_zone

See AvailabilityZone

blueprint

See Blueprint

bucket

See Bucket

bucket_bundle

See BucketBundle

bucket_state

See BucketState

bundle

See Bundle

cache_behavior

See CacheBehavior

cache_behavior_per_path

See CacheBehaviorPerPath

cache_settings

See CacheSettings

certificate

See Certificate

certificate_summary

See CertificateSummary

cloud_formation_stack_record

See CloudFormationStackRecord

cloud_formation_stack_record_source_info

See CloudFormationStackRecordSourceInfo

contact_method

See ContactMethod

container

See Container

container_image

See ContainerImage

container_service

See ContainerService

container_service_deployment

See ContainerServiceDeployment

container_service_deployment_request

See ContainerServiceDeploymentRequest

container_service_endpoint

See ContainerServiceEndpoint

container_service_health_check_config

See ContainerServiceHealthCheckConfig

container_service_log_event

See ContainerServiceLogEvent

container_service_power

See ContainerServicePower

container_service_registry_login

See ContainerServiceRegistryLogin

container_service_state_detail

See ContainerServiceStateDetail

cookie_object

See CookieObject

destination_info

See DestinationInfo

disk

See Disk

disk_info

See DiskInfo

disk_map

See DiskMap

disk_snapshot

See DiskSnapshot

disk_snapshot_info

See DiskSnapshotInfo

distribution_bundle

See DistributionBundle

domain

See Domain

domain_entry

See DomainEntry

domain_validation_record

See DomainValidationRecord

endpoint_request

See EndpointRequest

export_snapshot_record

See ExportSnapshotRecord

export_snapshot_record_source_info

See ExportSnapshotRecordSourceInfo

header_object

See HeaderObject

host_key_attributes

See HostKeyAttributes

input_origin

See InputOrigin

instance

See Instance

instance_access_details

See InstanceAccessDetails

instance_entry

See InstanceEntry

instance_hardware

See InstanceHardware

instance_health_summary

See InstanceHealthSummary

instance_networking

See InstanceNetworking

instance_port_info

See InstancePortInfo

instance_port_state

See InstancePortState

instance_snapshot

See InstanceSnapshot

instance_snapshot_info

See InstanceSnapshotInfo

instance_state

See InstanceState

key_pair

See KeyPair

lightsail_distribution

See LightsailDistribution

load_balancer

See LoadBalancer

load_balancer_tls_certificate

See LoadBalancerTlsCertificate

load_balancer_tls_certificate_domain_validation_option

See LoadBalancerTlsCertificateDomainValidationOption

load_balancer_tls_certificate_domain_validation_record

See LoadBalancerTlsCertificateDomainValidationRecord

load_balancer_tls_certificate_renewal_summary

See LoadBalancerTlsCertificateRenewalSummary

load_balancer_tls_certificate_summary

See LoadBalancerTlsCertificateSummary

log_event

See LogEvent

metric_datapoint

See MetricDatapoint

monitored_resource_info

See MonitoredResourceInfo

monthly_transfer

See MonthlyTransfer

operation

See Operation

origin

See Origin

password_data

See PasswordData

pending_maintenance_action

See PendingMaintenanceAction

pending_modified_relational_database_values

See PendingModifiedRelationalDatabaseValues

port_info

See PortInfo

query_string_object

See QueryStringObject

region

See Region

relational_database

See RelationalDatabase

relational_database_blueprint

See RelationalDatabaseBlueprint

relational_database_bundle

See RelationalDatabaseBundle

relational_database_endpoint

See RelationalDatabaseEndpoint

relational_database_event

See RelationalDatabaseEvent

relational_database_hardware

See RelationalDatabaseHardware

relational_database_parameter

See RelationalDatabaseParameter

relational_database_snapshot

See RelationalDatabaseSnapshot

renewal_summary

See RenewalSummary

resource_location

See ResourceLocation

resource_receiving_access

See ResourceReceivingAccess

resource_record

See ResourceRecord

static_ip

See StaticIp

tag

See Tag

Structs

AccessKey

Describes an access key for an Amazon Lightsail bucket.

AccessKeyLastUsed

Describes the last time an access key was used.

AccessRules

Describes the anonymous access permissions for an Amazon Lightsail bucket and its

AddOn

Describes an add-on that is enabled for an Amazon Lightsail resource.

AddOnRequest

Describes a request to enable, modify, or disable an add-on for an Amazon Lightsail

Alarm

Describes an alarm.

AttachedDisk

Describes a block storage disk that is attached to an instance, and is included in an

AutoSnapshotAddOnRequest

Describes a request to enable or modify the automatic snapshot add-on for an

AutoSnapshotDetails

Describes an automatic snapshot.

AvailabilityZone

Describes an Availability Zone.

Blueprint

Describes a blueprint (a virtual private server image).

Bucket

Describes an Amazon Lightsail bucket.

BucketBundle

Describes the specifications of a bundle that can be applied to an Amazon Lightsail

BucketState

Describes the state of an Amazon Lightsail bucket.

Bundle

Describes a bundle, which is a set of specs describing your virtual private server (or

CacheBehavior

Describes the default cache behavior of an Amazon Lightsail content delivery network (CDN)

CacheBehaviorPerPath

Describes the per-path cache behavior of an Amazon Lightsail content delivery network (CDN)

CacheSettings

Describes the cache settings of an Amazon Lightsail content delivery network (CDN)

Certificate

Describes the full details of an Amazon Lightsail SSL/TLS certificate.

CertificateSummary

Describes an Amazon Lightsail SSL/TLS certificate.

CloudFormationStackRecord

Describes a CloudFormation stack record created as a result of the create cloud

CloudFormationStackRecordSourceInfo

Describes the source of a CloudFormation stack record (i.e., the export snapshot

ContactMethod

Describes a contact method.

Container

Describes the settings of a container that will be launched, or that is launched, to an

ContainerImage

Describes a container image that is registered to an Amazon Lightsail container

ContainerService

Describes an Amazon Lightsail container service.

ContainerServiceDeployment

Describes a container deployment configuration of an Amazon Lightsail container

ContainerServiceDeploymentRequest

Describes a container deployment configuration of an Amazon Lightsail container

ContainerServiceEndpoint

Describes the public endpoint configuration of a deployment of an Amazon Lightsail

ContainerServiceHealthCheckConfig

Describes the health check configuration of an Amazon Lightsail container service.

ContainerServiceLogEvent

Describes the log events of a container of an Amazon Lightsail container service.

ContainerServicePower

Describes the powers that can be specified for an Amazon Lightsail container

ContainerServiceRegistryLogin

Describes the login information for the container image registry of an Amazon Lightsail

ContainerServiceStateDetail

Describes the current state of a container service.

CookieObject

Describes whether an Amazon Lightsail content delivery network (CDN) distribution forwards

DestinationInfo

Describes the destination of a record.

Disk

Describes a block storage disk.

DiskInfo

Describes a disk.

DiskMap

Describes a block storage disk mapping.

DiskSnapshot

Describes a block storage disk snapshot.

DiskSnapshotInfo

Describes a disk snapshot.

DistributionBundle

Describes the specifications of a distribution bundle.

Domain

Describes a domain where you are storing recordsets.

DomainEntry

Describes a domain recordset entry.

DomainValidationRecord

Describes the domain validation records of an Amazon Lightsail SSL/TLS certificate.

EndpointRequest

Describes the settings of a public endpoint for an Amazon Lightsail container

ExportSnapshotRecord

Describes an export snapshot record.

ExportSnapshotRecordSourceInfo

Describes the source of an export snapshot record.

HeaderObject

Describes the request headers that a Lightsail distribution bases caching on.

HostKeyAttributes

Describes the public SSH host keys or the RDP certificate.

InputOrigin

Describes the origin resource of an Amazon Lightsail content delivery network (CDN)

Instance

Describes an instance (a virtual private server).

InstanceAccessDetails

The parameters for gaining temporary access to one of your Amazon Lightsail

InstanceEntry

Describes the Amazon Elastic Compute Cloud instance and related resources to be created using the

InstanceHardware

Describes the hardware for the instance.

InstanceHealthSummary

Describes information about the health of the instance.

InstanceNetworking

Describes monthly data transfer rates and port information for an instance.

InstancePortInfo

Describes information about ports for an Amazon Lightsail instance.

InstancePortState

Describes open ports on an instance, the IP addresses allowed to connect to the instance

InstanceSnapshot

Describes an instance snapshot.

InstanceSnapshotInfo

Describes an instance snapshot.

InstanceState

Describes the virtual private server (or instance) status.

KeyPair

Describes an SSH key pair.

LightsailDistribution

Describes an Amazon Lightsail content delivery network (CDN) distribution.

LoadBalancer

Describes a load balancer.

LoadBalancerTlsCertificate

Describes a load balancer SSL/TLS certificate.

LoadBalancerTlsCertificateDomainValidationOption

Contains information about the domain names on an SSL/TLS certificate that you will use to

LoadBalancerTlsCertificateDomainValidationRecord

Describes the validation record of each domain name in the SSL/TLS certificate.

LoadBalancerTlsCertificateRenewalSummary

Contains information about the status of Lightsail's managed renewal for the

LoadBalancerTlsCertificateSummary

Provides a summary of SSL/TLS certificate metadata.

LogEvent

Describes a database log event.

MetricDatapoint

Describes the metric data point.

MonitoredResourceInfo

Describes resource being monitored by an alarm.

MonthlyTransfer

Describes the monthly data transfer in and out of your virtual private server (or

Operation

Describes the API operation.

Origin

Describes the origin resource of an Amazon Lightsail content delivery network (CDN)

PasswordData

The password data for the Windows Server-based instance, including the ciphertext and the

PendingMaintenanceAction

Describes a pending database maintenance action.

PendingModifiedRelationalDatabaseValues

Describes a pending database value modification.

PortInfo

Describes ports to open on an instance, the IP addresses allowed to connect to the

QueryStringObject

Describes the query string parameters that an Amazon Lightsail content delivery network

Region

Describes the AWS Region.

RelationalDatabase

Describes a database.

RelationalDatabaseBlueprint

Describes a database image, or blueprint. A blueprint describes the major engine version

RelationalDatabaseBundle

Describes a database bundle. A bundle describes the performance specifications of the

RelationalDatabaseEndpoint

Describes an endpoint for a database.

RelationalDatabaseEvent

Describes an event for a database.

RelationalDatabaseHardware

Describes the hardware of a database.

RelationalDatabaseParameter

Describes the parameters of a database.

RelationalDatabaseSnapshot

Describes a database snapshot.

RenewalSummary

Describes the status of a SSL/TLS certificate renewal managed by Amazon Lightsail.

ResourceLocation

Describes the resource location.

ResourceReceivingAccess

Describes an Amazon Lightsail instance that has access to a Lightsail bucket.

ResourceRecord

Describes the domain name system (DNS) records to add to your domain's DNS to validate it

StaticIp

Describes a static IP.

Tag

Describes a tag key and optional value assigned to an Amazon Lightsail resource.

Enums

AccessDirection

AccessType

AddOnType

AlarmState

AutoSnapshotStatus

BehaviorEnum

BlueprintType

BucketMetricName

CertificateStatus

CloudFormationStackRecordSourceType

ComparisonOperator

ContactMethodStatus

ContactMethodVerificationProtocol

ContactProtocol

ContainerServiceDeploymentState

ContainerServiceMetricName

ContainerServicePowerName

ContainerServiceProtocol

ContainerServiceState

ContainerServiceStateDetailCode

DiskSnapshotState

NOTE: DiskSnapshotState::Unknown has been renamed to ::UnknownValue.

DiskState

NOTE: DiskState::Unknown has been renamed to ::UnknownValue.

DistributionMetricName

ExportSnapshotRecordSourceType

ForwardValues

HeaderEnum

InstanceAccessProtocol

InstanceHealthReason

InstanceHealthState

InstanceMetricName

InstancePlatform

InstanceSnapshotState

IpAddressType

LoadBalancerAttributeName

LoadBalancerMetricName

LoadBalancerProtocol

LoadBalancerState

NOTE: LoadBalancerState::Unknown has been renamed to ::UnknownValue.

LoadBalancerTlsCertificateDomainStatus

LoadBalancerTlsCertificateFailureReason

LoadBalancerTlsCertificateRenewalStatus

LoadBalancerTlsCertificateRevocationReason

LoadBalancerTlsCertificateStatus

NOTE: LoadBalancerTlsCertificateStatus::Unknown has been renamed to ::UnknownValue.

MetricName

MetricStatistic

MetricUnit

NetworkProtocol

OperationStatus

OperationType

OriginProtocolPolicyEnum

PortAccessType

PortInfoSourceType

PortState

RecordState

RegionName

RelationalDatabaseEngine

RelationalDatabaseMetricName

RelationalDatabasePasswordVersion

RenewalStatus

ResourceBucketAccess

ResourceType

StatusType

TreatMissingData