Struct aws_sdk_lightsail::model::certificate::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Certificate

Implementations

impl Builder

pub fn arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the certificate.

pub fn set_arn(self, input: Option<String>) -> Self

pub fn name(self, input: impl Into<String>) -> Self

The name of the certificate (e.g., my-certificate).

pub fn set_name(self, input: Option<String>) -> Self

pub fn domain_name(self, input: impl Into<String>) -> Self

The domain name of the certificate.

pub fn set_domain_name(self, input: Option<String>) -> Self

pub fn status(self, input: CertificateStatus) -> Self

The validation status of the certificate.

pub fn set_status(self, input: Option<CertificateStatus>) -> Self

pub fn serial_number(self, input: impl Into<String>) -> Self

The serial number of the certificate.

pub fn set_serial_number(self, input: Option<String>) -> Self

pub fn subject_alternative_names(self, input: impl Into<String>) -> Self

pub fn set_subject_alternative_names(self, input: Option<Vec<String>>) -> Self

pub fn domain_validation_records(
self,
input: impl Into<DomainValidationRecord>
) -> Self

pub fn set_domain_validation_records(
self,
input: Option<Vec<DomainValidationRecord>>
) -> Self

pub fn request_failure_reason(self, input: impl Into<String>) -> Self

The validation failure reason, if any, of the certificate.

The following failure reasons are possible:

NO_AVAILABLE_CONTACTS - This failure applies to email validation, which is not available for Lightsail certificates.

ADDITIONAL_VERIFICATION_REQUIRED - Lightsail requires additional information to process this certificate request. This can happen as a fraud-protection measure, such as when the domain ranks within the Alexa top 1000 websites. To provide the required information, use the AWS Support Center to contact AWS Support.

You cannot request a certificate for Amazon-owned domain names such as those ending in amazonaws.com, cloudfront.net, or elasticbeanstalk.com.

DOMAIN_NOT_ALLOWED - One or more of the domain names in the certificate request was reported as an unsafe domain by VirusTotal. To correct the problem, search for your domain name on the VirusTotal website. If your domain is reported as suspicious, see Google Help for Hacked Websites to learn what you can do.

If you believe that the result is a false positive, notify the organization that is reporting the domain. VirusTotal is an aggregate of several antivirus and URL scanners and cannot remove your domain from a block list itself. After you correct the problem and the VirusTotal registry has been updated, request a new certificate.

If you see this error and your domain is not included in the VirusTotal list, visit the AWS Support Center and create a case.

INVALID_PUBLIC_DOMAIN - One or more of the domain names in the certificate request is not valid. Typically, this is because a domain name in the request is not a valid top-level domain. Try to request a certificate again, correcting any spelling errors or typos that were in the failed request, and ensure that all domain names in the request are for valid top-level domains. For example, you cannot request a certificate for example.invalidpublicdomain because invalidpublicdomain is not a valid top-level domain.

OTHER - Typically, this failure occurs when there is a typographical error in one or more of the domain names in the certificate request. Try to request a certificate again, correcting any spelling errors or typos that were in the failed request.

pub fn set_request_failure_reason(self, input: Option<String>) -> Self

pub fn in_use_resource_count(self, input: i32) -> Self

The number of Lightsail resources that the certificate is attached to.

pub fn set_in_use_resource_count(self, input: Option<i32>) -> Self

pub fn key_algorithm(self, input: impl Into<String>) -> Self

The algorithm used to generate the key pair (the public and private key) of the certificate.

pub fn set_key_algorithm(self, input: Option<String>) -> Self

pub fn created_at(self, input: Instant) -> Self

The timestamp when the certificate was created.

pub fn set_created_at(self, input: Option<Instant>) -> Self

pub fn issued_at(self, input: Instant) -> Self

The timestamp when the certificate was issued.

pub fn set_issued_at(self, input: Option<Instant>) -> Self

pub fn issuer_ca(self, input: impl Into<String>) -> Self

The certificate authority that issued the certificate.

pub fn set_issuer_ca(self, input: Option<String>) -> Self

pub fn not_before(self, input: Instant) -> Self

The timestamp when the certificate is first valid.

pub fn set_not_before(self, input: Option<Instant>) -> Self

pub fn not_after(self, input: Instant) -> Self

The timestamp when the certificate expires.

pub fn set_not_after(self, input: Option<Instant>) -> Self

pub fn eligible_to_renew(self, input: impl Into<String>) -> Self

The renewal eligibility of the certificate.

pub fn set_eligible_to_renew(self, input: Option<String>) -> Self

pub fn renewal_summary(self, input: RenewalSummary) -> Self

An object that describes the status of the certificate renewal managed by Lightsail.

pub fn set_renewal_summary(self, input: Option<RenewalSummary>) -> Self

pub fn revoked_at(self, input: Instant) -> Self

The timestamp when the certificate was revoked. This value is present only when the certificate status is REVOKED.

pub fn set_revoked_at(self, input: Option<Instant>) -> Self

pub fn revocation_reason(self, input: impl Into<String>) -> Self

The reason the certificate was revoked. This value is present only when the certificate status is REVOKED.

pub fn set_revocation_reason(self, input: Option<String>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn support_code(self, input: impl Into<String>) -> Self

The support code. Include this code in your email to support when you have questions about your Lightsail certificate. This code enables our support team to look up your Lightsail information more easily.

pub fn set_support_code(self, input: Option<String>) -> Self

pub fn build(self) -> Certificate

Consumes the builder and constructs a Certificate