Struct aws_sdk_lightsail::model::container_service::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for ContainerService

Implementations

impl Builder

pub fn container_service_name(self, input: impl Into<String>) -> Self

The name of the container service.

pub fn set_container_service_name(self, input: Option<String>) -> Self

pub fn arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the container service.

pub fn set_arn(self, input: Option<String>) -> Self

pub fn created_at(self, input: Instant) -> Self

The timestamp when the container service was created.

pub fn set_created_at(self, input: Option<Instant>) -> Self

pub fn location(self, input: ResourceLocation) -> Self

An object that describes the location of the container service, such as the AWS Region and Availability Zone.

pub fn set_location(self, input: Option<ResourceLocation>) -> Self

pub fn resource_type(self, input: ResourceType) -> Self

The Lightsail resource type of the container service (i.e., ContainerService).

pub fn set_resource_type(self, input: Option<ResourceType>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn power(self, input: ContainerServicePowerName) -> Self

The power specification of the container service.

The power specifies the amount of RAM, the number of vCPUs, and the base price of the container service.

pub fn set_power(self, input: Option<ContainerServicePowerName>) -> Self

pub fn power_id(self, input: impl Into<String>) -> Self

The ID of the power of the container service.

pub fn set_power_id(self, input: Option<String>) -> Self

pub fn state(self, input: ContainerServiceState) -> Self

The current state of the container service.

The following container service states are possible:

PENDING - The container service is being created.

READY - The container service is running but it does not have an active container deployment.

DEPLOYING - The container service is launching a container deployment.

RUNNING - The container service is running and it has an active container deployment.

UPDATING - The container service capacity or its custom domains are being updated.

DELETING - The container service is being deleted.

DISABLED - The container service is disabled, and its active deployment and containers, if any, are shut down.

pub fn set_state(self, input: Option<ContainerServiceState>) -> Self

pub fn state_detail(self, input: ContainerServiceStateDetail) -> Self

An object that describes the current state of the container service.

The state detail is populated only when a container service is in a PENDING, DEPLOYING, or UPDATING state.

pub fn set_state_detail(
self,
input: Option<ContainerServiceStateDetail>
) -> Self

pub fn scale(self, input: i32) -> Self

The scale specification of the container service.

The scale specifies the allocated compute nodes of the container service.

pub fn set_scale(self, input: Option<i32>) -> Self

pub fn current_deployment(self, input: ContainerServiceDeployment) -> Self

An object that describes the current container deployment of the container service.

pub fn set_current_deployment(
self,
input: Option<ContainerServiceDeployment>
) -> Self

pub fn next_deployment(self, input: ContainerServiceDeployment) -> Self

An object that describes the next deployment of the container service.

This value is null when there is no deployment in a pending state.

pub fn set_next_deployment(
self,
input: Option<ContainerServiceDeployment>
) -> Self

pub fn is_disabled(self, input: bool) -> Self

A Boolean value indicating whether the container service is disabled.

pub fn set_is_disabled(self, input: Option<bool>) -> Self

pub fn principal_arn(self, input: impl Into<String>) -> Self

The principal ARN of the container service.

The principal ARN can be used to create a trust relationship between your standard AWS account and your Lightsail container service. This allows you to give your service permission to access resources in your standard AWS account.

pub fn set_principal_arn(self, input: Option<String>) -> Self

pub fn private_domain_name(self, input: impl Into<String>) -> Self

The private domain name of the container service.

The private domain name is accessible only by other resources within the default virtual private cloud (VPC) of your Lightsail account.

pub fn set_private_domain_name(self, input: Option<String>) -> Self

pub fn public_domain_names(
self,
k: impl Into<String>,
v: impl Into<Vec<String>>
) -> Self

pub fn set_public_domain_names(
self,
input: Option<HashMap<String, Vec<String>>>
) -> Self

pub fn url(self, input: impl Into<String>) -> Self

The publicly accessible URL of the container service.

If no public endpoint is specified in the currentDeployment, this URL returns a 404 response.

pub fn set_url(self, input: Option<String>) -> Self

pub fn build(self) -> ContainerService

Consumes the builder and constructs a ContainerService