Struct aws_sdk_lightsail::model::Certificate

#[non_exhaustive]
pub struct Certificate {

    pub arn: Option<String>,
    pub name: Option<String>,
    pub domain_name: Option<String>,
    pub status: Option<CertificateStatus>,
    pub serial_number: Option<String>,
    pub subject_alternative_names: Option<Vec<String>>,
    pub domain_validation_records: Option<Vec<DomainValidationRecord>>,
    pub request_failure_reason: Option<String>,
    pub in_use_resource_count: i32,
    pub key_algorithm: Option<String>,
    pub created_at: Option<Instant>,
    pub issued_at: Option<Instant>,
    pub issuer_ca: Option<String>,
    pub not_before: Option<Instant>,
    pub not_after: Option<Instant>,
    pub eligible_to_renew: Option<String>,
    pub renewal_summary: Option<RenewalSummary>,
    pub revoked_at: Option<Instant>,
    pub revocation_reason: Option<String>,
    pub tags: Option<Vec<Tag>>,
    pub support_code: Option<String>,

}

Describes the full details of an Amazon Lightsail SSL/TLS certificate.

To get a summary of a certificate, use the GetCertificates action and ommit includeCertificateDetails from your request. The response will include only the certificate Amazon Resource Name (ARN), certificate name, domain name, and tags.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

arn: Option<String>

The Amazon Resource Name (ARN) of the certificate.

name: Option<String>

The name of the certificate (e.g., my-certificate).

domain_name: Option<String>

The domain name of the certificate.

status: Option<CertificateStatus>

The validation status of the certificate.

serial_number: Option<String>

The serial number of the certificate.

subject_alternative_names: Option<Vec<String>>

An array of strings that specify the alternate domains (e.g., example2.com) and subdomains (e.g., blog.example.com) of the certificate.

domain_validation_records: Option<Vec<DomainValidationRecord>>

An array of objects that describe the domain validation records of the certificate.

request_failure_reason: Option<String>

The validation failure reason, if any, of the certificate.

The following failure reasons are possible:

NO_AVAILABLE_CONTACTS - This failure applies to email validation, which is not available for Lightsail certificates.

ADDITIONAL_VERIFICATION_REQUIRED - Lightsail requires additional information to process this certificate request. This can happen as a fraud-protection measure, such as when the domain ranks within the Alexa top 1000 websites. To provide the required information, use the AWS Support Center to contact AWS Support.

You cannot request a certificate for Amazon-owned domain names such as those ending in amazonaws.com, cloudfront.net, or elasticbeanstalk.com.

DOMAIN_NOT_ALLOWED - One or more of the domain names in the certificate request was reported as an unsafe domain by VirusTotal. To correct the problem, search for your domain name on the VirusTotal website. If your domain is reported as suspicious, see Google Help for Hacked Websites to learn what you can do.

If you believe that the result is a false positive, notify the organization that is reporting the domain. VirusTotal is an aggregate of several antivirus and URL scanners and cannot remove your domain from a block list itself. After you correct the problem and the VirusTotal registry has been updated, request a new certificate.

If you see this error and your domain is not included in the VirusTotal list, visit the AWS Support Center and create a case.

INVALID_PUBLIC_DOMAIN - One or more of the domain names in the certificate request is not valid. Typically, this is because a domain name in the request is not a valid top-level domain. Try to request a certificate again, correcting any spelling errors or typos that were in the failed request, and ensure that all domain names in the request are for valid top-level domains. For example, you cannot request a certificate for example.invalidpublicdomain because invalidpublicdomain is not a valid top-level domain.

OTHER - Typically, this failure occurs when there is a typographical error in one or more of the domain names in the certificate request. Try to request a certificate again, correcting any spelling errors or typos that were in the failed request.

in_use_resource_count: i32

The number of Lightsail resources that the certificate is attached to.

key_algorithm: Option<String>

The algorithm used to generate the key pair (the public and private key) of the certificate.

created_at: Option<Instant>

The timestamp when the certificate was created.

issued_at: Option<Instant>

The timestamp when the certificate was issued.

issuer_ca: Option<String>

The certificate authority that issued the certificate.

not_before: Option<Instant>

The timestamp when the certificate is first valid.

not_after: Option<Instant>

The timestamp when the certificate expires.

eligible_to_renew: Option<String>

The renewal eligibility of the certificate.

renewal_summary: Option<RenewalSummary>

An object that describes the status of the certificate renewal managed by Lightsail.

revoked_at: Option<Instant>

The timestamp when the certificate was revoked. This value is present only when the certificate status is REVOKED.

revocation_reason: Option<String>

The reason the certificate was revoked. This value is present only when the certificate status is REVOKED.

tags: Option<Vec<Tag>>

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

support_code: Option<String>

The support code. Include this code in your email to support when you have questions about your Lightsail certificate. This code enables our support team to look up your Lightsail information more easily.

Implementations

impl Certificate

pub fn arn(&self) -> Option<&str>

The Amazon Resource Name (ARN) of the certificate.

pub fn name(&self) -> Option<&str>

The name of the certificate (e.g., my-certificate).

pub fn domain_name(&self) -> Option<&str>

The domain name of the certificate.

pub fn status(&self) -> Option<&CertificateStatus>

The validation status of the certificate.

pub fn serial_number(&self) -> Option<&str>

The serial number of the certificate.

pub fn subject_alternative_names(&self) -> Option<&[String]>

An array of strings that specify the alternate domains (e.g., example2.com) and subdomains (e.g., blog.example.com) of the certificate.

pub fn domain_validation_records(&self) -> Option<&[DomainValidationRecord]>

An array of objects that describe the domain validation records of the certificate.

pub fn request_failure_reason(&self) -> Option<&str>

The validation failure reason, if any, of the certificate.

The following failure reasons are possible:

NO_AVAILABLE_CONTACTS - This failure applies to email validation, which is not available for Lightsail certificates.

ADDITIONAL_VERIFICATION_REQUIRED - Lightsail requires additional information to process this certificate request. This can happen as a fraud-protection measure, such as when the domain ranks within the Alexa top 1000 websites. To provide the required information, use the AWS Support Center to contact AWS Support.

You cannot request a certificate for Amazon-owned domain names such as those ending in amazonaws.com, cloudfront.net, or elasticbeanstalk.com.

DOMAIN_NOT_ALLOWED - One or more of the domain names in the certificate request was reported as an unsafe domain by VirusTotal. To correct the problem, search for your domain name on the VirusTotal website. If your domain is reported as suspicious, see Google Help for Hacked Websites to learn what you can do.

If you believe that the result is a false positive, notify the organization that is reporting the domain. VirusTotal is an aggregate of several antivirus and URL scanners and cannot remove your domain from a block list itself. After you correct the problem and the VirusTotal registry has been updated, request a new certificate.

If you see this error and your domain is not included in the VirusTotal list, visit the AWS Support Center and create a case.

INVALID_PUBLIC_DOMAIN - One or more of the domain names in the certificate request is not valid. Typically, this is because a domain name in the request is not a valid top-level domain. Try to request a certificate again, correcting any spelling errors or typos that were in the failed request, and ensure that all domain names in the request are for valid top-level domains. For example, you cannot request a certificate for example.invalidpublicdomain because invalidpublicdomain is not a valid top-level domain.

OTHER - Typically, this failure occurs when there is a typographical error in one or more of the domain names in the certificate request. Try to request a certificate again, correcting any spelling errors or typos that were in the failed request.

pub fn in_use_resource_count(&self) -> i32

The number of Lightsail resources that the certificate is attached to.

pub fn key_algorithm(&self) -> Option<&str>

The algorithm used to generate the key pair (the public and private key) of the certificate.

pub fn created_at(&self) -> Option<&Instant>

The timestamp when the certificate was created.

pub fn issued_at(&self) -> Option<&Instant>

The timestamp when the certificate was issued.

pub fn issuer_ca(&self) -> Option<&str>

The certificate authority that issued the certificate.

pub fn not_before(&self) -> Option<&Instant>

The timestamp when the certificate is first valid.

pub fn not_after(&self) -> Option<&Instant>

The timestamp when the certificate expires.

pub fn eligible_to_renew(&self) -> Option<&str>

The renewal eligibility of the certificate.

pub fn renewal_summary(&self) -> Option<&RenewalSummary>

An object that describes the status of the certificate renewal managed by Lightsail.

pub fn revoked_at(&self) -> Option<&Instant>

The timestamp when the certificate was revoked. This value is present only when the certificate status is REVOKED.

pub fn revocation_reason(&self) -> Option<&str>

The reason the certificate was revoked. This value is present only when the certificate status is REVOKED.

pub fn tags(&self) -> Option<&[Tag]>

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

pub fn support_code(&self) -> Option<&str>

The support code. Include this code in your email to support when you have questions about your Lightsail certificate. This code enables our support team to look up your Lightsail information more easily.

impl Certificate

pub fn builder() -> Builder

Creates a new builder-style object to manufacture Certificate