Module aws_sdk_lightsail::operation

Structs

AllocateStaticIp

Allocates a static IP address.

AttachCertificateToDistribution

Attaches an SSL/TLS certificate to your Amazon Lightsail content delivery network (CDN)

AttachDisk

Attaches a block storage disk to a running or stopped Lightsail instance and exposes it

AttachInstancesToLoadBalancer

Attaches one or more Lightsail instances to a load balancer.

AttachLoadBalancerTlsCertificate

Attaches a Transport Layer Security (TLS) certificate to your load balancer. TLS is just

AttachStaticIp

Attaches a static IP address to a specific Amazon Lightsail instance.

CloseInstancePublicPorts

Closes ports for a specific Amazon Lightsail instance.

CopySnapshot

Copies a manual snapshot of an instance or disk as another manual snapshot, or copies an

CreateBucket

Creates an Amazon Lightsail bucket.

CreateBucketAccessKey

Creates a new access key for the specified Amazon Lightsail bucket. Access keys consist of

CreateCertificate

Creates an SSL/TLS certificate for an Amazon Lightsail content delivery network (CDN)

CreateCloudFormationStack

Creates an AWS CloudFormation stack, which creates a new Amazon EC2 instance from an exported

CreateContactMethod

Creates an email or SMS text message contact method.

CreateContainerService

Creates an Amazon Lightsail container service.

CreateContainerServiceDeployment

Creates a deployment for your Amazon Lightsail container service.

CreateContainerServiceRegistryLogin

Creates a temporary set of log in credentials that you can use to log in to the Docker

CreateDisk

Creates a block storage disk that can be attached to an Amazon Lightsail instance in the

CreateDiskFromSnapshot

Creates a block storage disk from a manual or automatic snapshot of a disk. The resulting

CreateDiskSnapshot

Creates a snapshot of a block storage disk. You can use snapshots for backups, to make

CreateDistribution

Creates an Amazon Lightsail content delivery network (CDN) distribution.

CreateDomain

Creates a domain resource for the specified domain (e.g., example.com).

CreateDomainEntry

Creates one of the following domain name system (DNS) records in a domain DNS zone:

CreateInstanceSnapshot

Creates a snapshot of a specific virtual private server, or instance.

CreateInstances

Creates one or more Amazon Lightsail instances.

CreateInstancesFromSnapshot

Creates one or more new instances from a manual or automatic snapshot of an

CreateKeyPair

Creates an SSH key pair.

CreateLoadBalancer

Creates a Lightsail load balancer. To learn more about deciding whether to load balance

CreateLoadBalancerTlsCertificate

Creates an SSL/TLS certificate for an Amazon Lightsail load balancer.

CreateRelationalDatabase

Creates a new database in Amazon Lightsail.

CreateRelationalDatabaseFromSnapshot

Creates a new database from an existing database snapshot in Amazon Lightsail.

CreateRelationalDatabaseSnapshot

Creates a snapshot of your database in Amazon Lightsail. You can use snapshots for backups,

DeleteAlarm

Deletes an alarm.

DeleteAutoSnapshot

Deletes an automatic snapshot of an instance or disk. For more information, see the Amazon Lightsail Developer Guide.

DeleteBucket

Deletes a Amazon Lightsail bucket.

DeleteBucketAccessKey

Deletes an access key for the specified Amazon Lightsail bucket.

DeleteCertificate

Deletes an SSL/TLS certificate for your Amazon Lightsail content delivery network (CDN)

DeleteContactMethod

Deletes a contact method.

DeleteContainerImage

Deletes a container image that is registered to your Amazon Lightsail container

DeleteContainerService

Deletes your Amazon Lightsail container service.

DeleteDisk

Deletes the specified block storage disk. The disk must be in the available

DeleteDiskSnapshot

Deletes the specified disk snapshot.

DeleteDistribution

Deletes your Amazon Lightsail content delivery network (CDN) distribution.

DeleteDomain

Deletes the specified domain recordset and all of its domain records.

DeleteDomainEntry

Deletes a specific domain entry.

DeleteInstance

Deletes an Amazon Lightsail instance.

DeleteInstanceSnapshot

Deletes a specific snapshot of a virtual private server (or

DeleteKeyPair

Deletes a specific SSH key pair.

DeleteKnownHostKeys

Deletes the known host key or certificate used by the Amazon Lightsail browser-based SSH or

DeleteLoadBalancer

Deletes a Lightsail load balancer and all its associated SSL/TLS certificates. Once the

DeleteLoadBalancerTlsCertificate

Deletes an SSL/TLS certificate associated with a Lightsail load balancer.

DeleteRelationalDatabase

Deletes a database in Amazon Lightsail.

DeleteRelationalDatabaseSnapshot

Deletes a database snapshot in Amazon Lightsail.

DetachCertificateFromDistribution

Detaches an SSL/TLS certificate from your Amazon Lightsail content delivery network (CDN)

DetachDisk

Detaches a stopped block storage disk from a Lightsail instance. Make sure to unmount

DetachInstancesFromLoadBalancer

Detaches the specified instances from a Lightsail load balancer.

DetachStaticIp

Detaches a static IP from the Amazon Lightsail instance to which it is attached.

DisableAddOn

Disables an add-on for an Amazon Lightsail resource. For more information, see the Amazon Lightsail Developer Guide.

DownloadDefaultKeyPair

Downloads the default SSH key pair from the user's account.

EnableAddOn

Enables or modifies an add-on for an Amazon Lightsail resource. For more information, see

ExportSnapshot

Exports an Amazon Lightsail instance or block storage disk snapshot to Amazon Elastic Compute Cloud (Amazon EC2).

GetActiveNames

Returns the names of all active (not deleted) resources.

GetAlarms

Returns information about the configured alarms. Specify an alarm name in your request to

GetAutoSnapshots

Returns the available automatic snapshots for an instance or disk. For more information,

GetBlueprints

Returns the list of available instance images, or blueprints. You can

GetBucketAccessKeys

Returns the existing access key IDs for the specified Amazon Lightsail bucket.

GetBucketBundles

Returns the bundles that you can apply to a Amazon Lightsail bucket.

GetBucketMetricData

Returns the data points of a specific metric for an Amazon Lightsail bucket.

GetBuckets

Returns information about one or more Amazon Lightsail buckets.

GetBundles

Returns the list of bundles that are available for purchase. A bundle describes the specs

GetCertificates

Returns information about one or more Amazon Lightsail SSL/TLS certificates.

GetCloudFormationStackRecords

Returns the CloudFormation stack record created as a result of the create cloud

GetContactMethods

Returns information about the configured contact methods. Specify a protocol in your

GetContainerAPIMetadata

Returns information about Amazon Lightsail containers, such as the current version of the

GetContainerImages

Returns the container images that are registered to your Amazon Lightsail container

GetContainerLog

Returns the log events of a container of your Amazon Lightsail container service.

GetContainerServiceDeployments

Returns the deployments for your Amazon Lightsail container service

GetContainerServiceMetricData

Returns the data points of a specific metric of your Amazon Lightsail container

GetContainerServicePowers

Returns the list of powers that can be specified for your Amazon Lightsail container

GetContainerServices

Returns information about one or more of your Amazon Lightsail container services.

GetDisk

Returns information about a specific block storage disk.

GetDiskSnapshot

Returns information about a specific block storage disk snapshot.

GetDiskSnapshots

Returns information about all block storage disk snapshots in your AWS account and

GetDisks

Returns information about all block storage disks in your AWS account and region.

GetDistributionBundles

Returns the bundles that can be applied to your Amazon Lightsail content delivery network

GetDistributionLatestCacheReset

Returns the timestamp and status of the last cache reset of a specific Amazon Lightsail

GetDistributionMetricData

Returns the data points of a specific metric for an Amazon Lightsail content delivery

GetDistributions

Returns information about one or more of your Amazon Lightsail content delivery network

GetDomain

Returns information about a specific domain recordset.

GetDomains

Returns a list of all domains in the user's account.

GetExportSnapshotRecords

Returns all export snapshot records created as a result of the export

GetInstance

Returns information about a specific Amazon Lightsail instance, which is a virtual private

GetInstanceAccessDetails

Returns temporary SSH keys you can use to connect to a specific virtual private server, or

GetInstanceMetricData

Returns the data points for the specified Amazon Lightsail instance metric, given an

GetInstancePortStates

Returns the firewall port states for a specific Amazon Lightsail instance, the IP addresses

GetInstanceSnapshot

Returns information about a specific instance snapshot.

GetInstanceSnapshots

Returns all instance snapshots for the user's account.

GetInstanceState

Returns the state of a specific instance. Works on one instance at a time.

GetInstances

Returns information about all Amazon Lightsail virtual private servers, or

GetKeyPair

Returns information about a specific key pair.

GetKeyPairs

Returns information about all key pairs in the user's account.

GetLoadBalancer

Returns information about the specified Lightsail load balancer.

GetLoadBalancerMetricData

Returns information about health metrics for your Lightsail load balancer.

GetLoadBalancerTlsCertificates

Returns information about the TLS certificates that are associated with the specified

GetLoadBalancers

Returns information about all load balancers in an account.

GetOperation

Returns information about a specific operation. Operations include events such as when you

GetOperations

Returns information about all operations.

GetOperationsForResource

Gets operations for a specific resource (e.g., an instance or a static IP).

GetRegions

Returns a list of all valid regions for Amazon Lightsail. Use the include

GetRelationalDatabase

Returns information about a specific database in Amazon Lightsail.

GetRelationalDatabaseBlueprints

Returns a list of available database blueprints in Amazon Lightsail. A blueprint describes

GetRelationalDatabaseBundles

Returns the list of bundles that are available in Amazon Lightsail. A bundle describes the

GetRelationalDatabaseEvents

Returns a list of events for a specific database in Amazon Lightsail.

GetRelationalDatabaseLogEvents

Returns a list of log events for a database in Amazon Lightsail.

GetRelationalDatabaseLogStreams

Returns a list of available log streams for a specific database in Amazon Lightsail.

GetRelationalDatabaseMasterUserPassword

Returns the current, previous, or pending versions of the master user password for a

GetRelationalDatabaseMetricData

Returns the data points of the specified metric for a database in Amazon Lightsail.

GetRelationalDatabaseParameters

Returns all of the runtime parameters offered by the underlying database software, or

GetRelationalDatabaseSnapshot

Returns information about a specific database snapshot in Amazon Lightsail.

GetRelationalDatabaseSnapshots

Returns information about all of your database snapshots in Amazon Lightsail.

GetRelationalDatabases

Returns information about all of your databases in Amazon Lightsail.

GetStaticIp

Returns information about an Amazon Lightsail static IP.

GetStaticIps

Returns information about all static IPs in the user's account.

ImportKeyPair

Imports a public SSH key from a specific key pair.

IsVpcPeered

Returns a Boolean value indicating whether your Lightsail VPC is peered.

OpenInstancePublicPorts

Opens ports for a specific Amazon Lightsail instance, and specifies the IP addresses

PeerVpc

Peers the Lightsail VPC with the user's default VPC.

PutAlarm

Creates or updates an alarm, and associates it with the specified metric.

PutInstancePublicPorts

Opens ports for a specific Amazon Lightsail instance, and specifies the IP addresses

RebootInstance

Restarts a specific instance.

RebootRelationalDatabase

Restarts a specific database in Amazon Lightsail.

RegisterContainerImage

Registers a container image to your Amazon Lightsail container service.

ReleaseStaticIp

Deletes a specific static IP from your account.

ResetDistributionCache

Deletes currently cached content from your Amazon Lightsail content delivery network (CDN)

SendContactMethodVerification

Sends a verification request to an email contact method to ensure it's owned by the

SetIpAddressType

Sets the IP address type for an Amazon Lightsail resource.

SetResourceAccessForBucket

Sets the Amazon Lightsail resources that can access the specified Lightsail

StartInstance

Starts a specific Amazon Lightsail instance from a stopped state. To restart an instance,

StartRelationalDatabase

Starts a specific database from a stopped state in Amazon Lightsail. To restart a database,

StopInstance

Stops a specific Amazon Lightsail instance that is currently running.

StopRelationalDatabase

Stops a specific database that is currently running in Amazon Lightsail.

TagResource

Adds one or more tags to the specified Amazon Lightsail resource. Each resource can have a

TestAlarm

Tests an alarm by displaying a banner on the Amazon Lightsail console. If a notification

UnpeerVpc

Unpeers the Lightsail VPC from the user's default VPC.

UntagResource

Deletes the specified set of tag keys and their values from the specified Amazon Lightsail

UpdateBucket

Updates an existing Amazon Lightsail bucket.

UpdateBucketBundle

Updates the bundle, or storage plan, of an existing Amazon Lightsail bucket.

UpdateContainerService

Updates the configuration of your Amazon Lightsail container service, such as its power,

UpdateDistribution

Updates an existing Amazon Lightsail content delivery network (CDN) distribution.

UpdateDistributionBundle

Updates the bundle of your Amazon Lightsail content delivery network (CDN)

UpdateDomainEntry

Updates a domain recordset after it is created.

UpdateLoadBalancerAttribute

Updates the specified attribute for a load balancer. You can only update one attribute at

UpdateRelationalDatabase

Allows the update of one or more attributes of a database in Amazon Lightsail.

UpdateRelationalDatabaseParameters

Allows the update of one or more parameters of a database in Amazon Lightsail.