Struct aws_sdk_lightsail::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for Lightsail_20161128.

This client allows ergonomic access to a Lightsail_20161128-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn allocate_static_ip(&self) -> AllocateStaticIp<C, M, R>

Constructs a fluent builder for the AllocateStaticIp operation.

See AllocateStaticIp for more information about the operation and its arguments.

pub fn attach_certificate_to_distribution(
&self
) -> AttachCertificateToDistribution<C, M, R>

Constructs a fluent builder for the AttachCertificateToDistribution operation.

See AttachCertificateToDistribution for more information about the operation and its arguments.

pub fn attach_disk(&self) -> AttachDisk<C, M, R>

Constructs a fluent builder for the AttachDisk operation.

See AttachDisk for more information about the operation and its arguments.

pub fn attach_instances_to_load_balancer(
&self
) -> AttachInstancesToLoadBalancer<C, M, R>

Constructs a fluent builder for the AttachInstancesToLoadBalancer operation.

See AttachInstancesToLoadBalancer for more information about the operation and its arguments.

pub fn attach_load_balancer_tls_certificate(
&self
) -> AttachLoadBalancerTlsCertificate<C, M, R>

Constructs a fluent builder for the AttachLoadBalancerTlsCertificate operation.

See AttachLoadBalancerTlsCertificate for more information about the operation and its arguments.

pub fn attach_static_ip(&self) -> AttachStaticIp<C, M, R>

Constructs a fluent builder for the AttachStaticIp operation.

See AttachStaticIp for more information about the operation and its arguments.

pub fn close_instance_public_ports(&self) -> CloseInstancePublicPorts<C, M, R>

Constructs a fluent builder for the CloseInstancePublicPorts operation.

See CloseInstancePublicPorts for more information about the operation and its arguments.

pub fn copy_snapshot(&self) -> CopySnapshot<C, M, R>

Constructs a fluent builder for the CopySnapshot operation.

See CopySnapshot for more information about the operation and its arguments.

pub fn create_bucket(&self) -> CreateBucket<C, M, R>

Constructs a fluent builder for the CreateBucket operation.

See CreateBucket for more information about the operation and its arguments.

pub fn create_bucket_access_key(&self) -> CreateBucketAccessKey<C, M, R>

Constructs a fluent builder for the CreateBucketAccessKey operation.

See CreateBucketAccessKey for more information about the operation and its arguments.

pub fn create_certificate(&self) -> CreateCertificate<C, M, R>

Constructs a fluent builder for the CreateCertificate operation.

See CreateCertificate for more information about the operation and its arguments.

pub fn create_cloud_formation_stack(&self) -> CreateCloudFormationStack<C, M, R>

Constructs a fluent builder for the CreateCloudFormationStack operation.

See CreateCloudFormationStack for more information about the operation and its arguments.

pub fn create_contact_method(&self) -> CreateContactMethod<C, M, R>

Constructs a fluent builder for the CreateContactMethod operation.

See CreateContactMethod for more information about the operation and its arguments.

pub fn create_container_service(&self) -> CreateContainerService<C, M, R>

Constructs a fluent builder for the CreateContainerService operation.

See CreateContainerService for more information about the operation and its arguments.

pub fn create_container_service_deployment(
&self
) -> CreateContainerServiceDeployment<C, M, R>

Constructs a fluent builder for the CreateContainerServiceDeployment operation.

See CreateContainerServiceDeployment for more information about the operation and its arguments.

pub fn create_container_service_registry_login(
&self
) -> CreateContainerServiceRegistryLogin<C, M, R>

Constructs a fluent builder for the CreateContainerServiceRegistryLogin operation.

See CreateContainerServiceRegistryLogin for more information about the operation and its arguments.

pub fn create_disk(&self) -> CreateDisk<C, M, R>

Constructs a fluent builder for the CreateDisk operation.

See CreateDisk for more information about the operation and its arguments.

pub fn create_disk_from_snapshot(&self) -> CreateDiskFromSnapshot<C, M, R>

Constructs a fluent builder for the CreateDiskFromSnapshot operation.

See CreateDiskFromSnapshot for more information about the operation and its arguments.

pub fn create_disk_snapshot(&self) -> CreateDiskSnapshot<C, M, R>

Constructs a fluent builder for the CreateDiskSnapshot operation.

See CreateDiskSnapshot for more information about the operation and its arguments.

pub fn create_distribution(&self) -> CreateDistribution<C, M, R>

Constructs a fluent builder for the CreateDistribution operation.

See CreateDistribution for more information about the operation and its arguments.

pub fn create_domain(&self) -> CreateDomain<C, M, R>

Constructs a fluent builder for the CreateDomain operation.

See CreateDomain for more information about the operation and its arguments.

pub fn create_domain_entry(&self) -> CreateDomainEntry<C, M, R>

Constructs a fluent builder for the CreateDomainEntry operation.

See CreateDomainEntry for more information about the operation and its arguments.

pub fn create_instances(&self) -> CreateInstances<C, M, R>

Constructs a fluent builder for the CreateInstances operation.

See CreateInstances for more information about the operation and its arguments.

pub fn create_instances_from_snapshot(
&self
) -> CreateInstancesFromSnapshot<C, M, R>

Constructs a fluent builder for the CreateInstancesFromSnapshot operation.

See CreateInstancesFromSnapshot for more information about the operation and its arguments.

pub fn create_instance_snapshot(&self) -> CreateInstanceSnapshot<C, M, R>

Constructs a fluent builder for the CreateInstanceSnapshot operation.

See CreateInstanceSnapshot for more information about the operation and its arguments.

pub fn create_key_pair(&self) -> CreateKeyPair<C, M, R>

Constructs a fluent builder for the CreateKeyPair operation.

See CreateKeyPair for more information about the operation and its arguments.

pub fn create_load_balancer(&self) -> CreateLoadBalancer<C, M, R>

Constructs a fluent builder for the CreateLoadBalancer operation.

See CreateLoadBalancer for more information about the operation and its arguments.

pub fn create_load_balancer_tls_certificate(
&self
) -> CreateLoadBalancerTlsCertificate<C, M, R>

Constructs a fluent builder for the CreateLoadBalancerTlsCertificate operation.

See CreateLoadBalancerTlsCertificate for more information about the operation and its arguments.

pub fn create_relational_database(&self) -> CreateRelationalDatabase<C, M, R>

Constructs a fluent builder for the CreateRelationalDatabase operation.

See CreateRelationalDatabase for more information about the operation and its arguments.

pub fn create_relational_database_from_snapshot(
&self
) -> CreateRelationalDatabaseFromSnapshot<C, M, R>

Constructs a fluent builder for the CreateRelationalDatabaseFromSnapshot operation.

See CreateRelationalDatabaseFromSnapshot for more information about the operation and its arguments.

pub fn create_relational_database_snapshot(
&self
) -> CreateRelationalDatabaseSnapshot<C, M, R>

Constructs a fluent builder for the CreateRelationalDatabaseSnapshot operation.

See CreateRelationalDatabaseSnapshot for more information about the operation and its arguments.

pub fn delete_alarm(&self) -> DeleteAlarm<C, M, R>

Constructs a fluent builder for the DeleteAlarm operation.

See DeleteAlarm for more information about the operation and its arguments.

pub fn delete_auto_snapshot(&self) -> DeleteAutoSnapshot<C, M, R>

Constructs a fluent builder for the DeleteAutoSnapshot operation.

See DeleteAutoSnapshot for more information about the operation and its arguments.

pub fn delete_bucket(&self) -> DeleteBucket<C, M, R>

Constructs a fluent builder for the DeleteBucket operation.

See DeleteBucket for more information about the operation and its arguments.

pub fn delete_bucket_access_key(&self) -> DeleteBucketAccessKey<C, M, R>

Constructs a fluent builder for the DeleteBucketAccessKey operation.

See DeleteBucketAccessKey for more information about the operation and its arguments.

pub fn delete_certificate(&self) -> DeleteCertificate<C, M, R>

Constructs a fluent builder for the DeleteCertificate operation.

See DeleteCertificate for more information about the operation and its arguments.

pub fn delete_contact_method(&self) -> DeleteContactMethod<C, M, R>

Constructs a fluent builder for the DeleteContactMethod operation.

See DeleteContactMethod for more information about the operation and its arguments.

pub fn delete_container_image(&self) -> DeleteContainerImage<C, M, R>

Constructs a fluent builder for the DeleteContainerImage operation.

See DeleteContainerImage for more information about the operation and its arguments.

pub fn delete_container_service(&self) -> DeleteContainerService<C, M, R>

Constructs a fluent builder for the DeleteContainerService operation.

See DeleteContainerService for more information about the operation and its arguments.

pub fn delete_disk(&self) -> DeleteDisk<C, M, R>

Constructs a fluent builder for the DeleteDisk operation.

See DeleteDisk for more information about the operation and its arguments.

pub fn delete_disk_snapshot(&self) -> DeleteDiskSnapshot<C, M, R>

Constructs a fluent builder for the DeleteDiskSnapshot operation.

See DeleteDiskSnapshot for more information about the operation and its arguments.

pub fn delete_distribution(&self) -> DeleteDistribution<C, M, R>

Constructs a fluent builder for the DeleteDistribution operation.

See DeleteDistribution for more information about the operation and its arguments.

pub fn delete_domain(&self) -> DeleteDomain<C, M, R>

Constructs a fluent builder for the DeleteDomain operation.

See DeleteDomain for more information about the operation and its arguments.

pub fn delete_domain_entry(&self) -> DeleteDomainEntry<C, M, R>

Constructs a fluent builder for the DeleteDomainEntry operation.

See DeleteDomainEntry for more information about the operation and its arguments.

pub fn delete_instance(&self) -> DeleteInstance<C, M, R>

Constructs a fluent builder for the DeleteInstance operation.

See DeleteInstance for more information about the operation and its arguments.

pub fn delete_instance_snapshot(&self) -> DeleteInstanceSnapshot<C, M, R>

Constructs a fluent builder for the DeleteInstanceSnapshot operation.

See DeleteInstanceSnapshot for more information about the operation and its arguments.

pub fn delete_key_pair(&self) -> DeleteKeyPair<C, M, R>

Constructs a fluent builder for the DeleteKeyPair operation.

See DeleteKeyPair for more information about the operation and its arguments.

pub fn delete_known_host_keys(&self) -> DeleteKnownHostKeys<C, M, R>

Constructs a fluent builder for the DeleteKnownHostKeys operation.

See DeleteKnownHostKeys for more information about the operation and its arguments.

pub fn delete_load_balancer(&self) -> DeleteLoadBalancer<C, M, R>

Constructs a fluent builder for the DeleteLoadBalancer operation.

See DeleteLoadBalancer for more information about the operation and its arguments.

pub fn delete_load_balancer_tls_certificate(
&self
) -> DeleteLoadBalancerTlsCertificate<C, M, R>

Constructs a fluent builder for the DeleteLoadBalancerTlsCertificate operation.

See DeleteLoadBalancerTlsCertificate for more information about the operation and its arguments.

pub fn delete_relational_database(&self) -> DeleteRelationalDatabase<C, M, R>

Constructs a fluent builder for the DeleteRelationalDatabase operation.

See DeleteRelationalDatabase for more information about the operation and its arguments.

pub fn delete_relational_database_snapshot(
&self
) -> DeleteRelationalDatabaseSnapshot<C, M, R>

Constructs a fluent builder for the DeleteRelationalDatabaseSnapshot operation.

See DeleteRelationalDatabaseSnapshot for more information about the operation and its arguments.

pub fn detach_certificate_from_distribution(
&self
) -> DetachCertificateFromDistribution<C, M, R>

Constructs a fluent builder for the DetachCertificateFromDistribution operation.

See DetachCertificateFromDistribution for more information about the operation and its arguments.

pub fn detach_disk(&self) -> DetachDisk<C, M, R>

Constructs a fluent builder for the DetachDisk operation.

See DetachDisk for more information about the operation and its arguments.

pub fn detach_instances_from_load_balancer(
&self
) -> DetachInstancesFromLoadBalancer<C, M, R>

Constructs a fluent builder for the DetachInstancesFromLoadBalancer operation.

See DetachInstancesFromLoadBalancer for more information about the operation and its arguments.

pub fn detach_static_ip(&self) -> DetachStaticIp<C, M, R>

Constructs a fluent builder for the DetachStaticIp operation.

See DetachStaticIp for more information about the operation and its arguments.

pub fn disable_add_on(&self) -> DisableAddOn<C, M, R>

Constructs a fluent builder for the DisableAddOn operation.

See DisableAddOn for more information about the operation and its arguments.

pub fn download_default_key_pair(&self) -> DownloadDefaultKeyPair<C, M, R>

Constructs a fluent builder for the DownloadDefaultKeyPair operation.

See DownloadDefaultKeyPair for more information about the operation and its arguments.

pub fn enable_add_on(&self) -> EnableAddOn<C, M, R>

Constructs a fluent builder for the EnableAddOn operation.

See EnableAddOn for more information about the operation and its arguments.

pub fn export_snapshot(&self) -> ExportSnapshot<C, M, R>

Constructs a fluent builder for the ExportSnapshot operation.

See ExportSnapshot for more information about the operation and its arguments.

pub fn get_active_names(&self) -> GetActiveNames<C, M, R>

Constructs a fluent builder for the GetActiveNames operation.

See GetActiveNames for more information about the operation and its arguments.

pub fn get_alarms(&self) -> GetAlarms<C, M, R>

Constructs a fluent builder for the GetAlarms operation.

See GetAlarms for more information about the operation and its arguments.

pub fn get_auto_snapshots(&self) -> GetAutoSnapshots<C, M, R>

Constructs a fluent builder for the GetAutoSnapshots operation.

See GetAutoSnapshots for more information about the operation and its arguments.

pub fn get_blueprints(&self) -> GetBlueprints<C, M, R>

Constructs a fluent builder for the GetBlueprints operation.

See GetBlueprints for more information about the operation and its arguments.

pub fn get_bucket_access_keys(&self) -> GetBucketAccessKeys<C, M, R>

Constructs a fluent builder for the GetBucketAccessKeys operation.

See GetBucketAccessKeys for more information about the operation and its arguments.

pub fn get_bucket_bundles(&self) -> GetBucketBundles<C, M, R>

Constructs a fluent builder for the GetBucketBundles operation.

See GetBucketBundles for more information about the operation and its arguments.

pub fn get_bucket_metric_data(&self) -> GetBucketMetricData<C, M, R>

Constructs a fluent builder for the GetBucketMetricData operation.

See GetBucketMetricData for more information about the operation and its arguments.

pub fn get_buckets(&self) -> GetBuckets<C, M, R>

Constructs a fluent builder for the GetBuckets operation.

See GetBuckets for more information about the operation and its arguments.

pub fn get_bundles(&self) -> GetBundles<C, M, R>

Constructs a fluent builder for the GetBundles operation.

See GetBundles for more information about the operation and its arguments.

pub fn get_certificates(&self) -> GetCertificates<C, M, R>

Constructs a fluent builder for the GetCertificates operation.

See GetCertificates for more information about the operation and its arguments.

pub fn get_cloud_formation_stack_records(
&self
) -> GetCloudFormationStackRecords<C, M, R>

Constructs a fluent builder for the GetCloudFormationStackRecords operation.

See GetCloudFormationStackRecords for more information about the operation and its arguments.

pub fn get_contact_methods(&self) -> GetContactMethods<C, M, R>

Constructs a fluent builder for the GetContactMethods operation.

See GetContactMethods for more information about the operation and its arguments.

pub fn get_container_api_metadata(&self) -> GetContainerAPIMetadata<C, M, R>

Constructs a fluent builder for the GetContainerAPIMetadata operation.

See GetContainerAPIMetadata for more information about the operation and its arguments.

pub fn get_container_images(&self) -> GetContainerImages<C, M, R>

Constructs a fluent builder for the GetContainerImages operation.

See GetContainerImages for more information about the operation and its arguments.

pub fn get_container_log(&self) -> GetContainerLog<C, M, R>

Constructs a fluent builder for the GetContainerLog operation.

See GetContainerLog for more information about the operation and its arguments.

pub fn get_container_service_deployments(
&self
) -> GetContainerServiceDeployments<C, M, R>

Constructs a fluent builder for the GetContainerServiceDeployments operation.

See GetContainerServiceDeployments for more information about the operation and its arguments.

pub fn get_container_service_metric_data(
&self
) -> GetContainerServiceMetricData<C, M, R>

Constructs a fluent builder for the GetContainerServiceMetricData operation.

See GetContainerServiceMetricData for more information about the operation and its arguments.

pub fn get_container_service_powers(&self) -> GetContainerServicePowers<C, M, R>

Constructs a fluent builder for the GetContainerServicePowers operation.

See GetContainerServicePowers for more information about the operation and its arguments.

pub fn get_container_services(&self) -> GetContainerServices<C, M, R>

Constructs a fluent builder for the GetContainerServices operation.

See GetContainerServices for more information about the operation and its arguments.

pub fn get_disk(&self) -> GetDisk<C, M, R>

Constructs a fluent builder for the GetDisk operation.

See GetDisk for more information about the operation and its arguments.

pub fn get_disks(&self) -> GetDisks<C, M, R>

Constructs a fluent builder for the GetDisks operation.

See GetDisks for more information about the operation and its arguments.

pub fn get_disk_snapshot(&self) -> GetDiskSnapshot<C, M, R>

Constructs a fluent builder for the GetDiskSnapshot operation.

See GetDiskSnapshot for more information about the operation and its arguments.

pub fn get_disk_snapshots(&self) -> GetDiskSnapshots<C, M, R>

Constructs a fluent builder for the GetDiskSnapshots operation.

See GetDiskSnapshots for more information about the operation and its arguments.

pub fn get_distribution_bundles(&self) -> GetDistributionBundles<C, M, R>

Constructs a fluent builder for the GetDistributionBundles operation.

See GetDistributionBundles for more information about the operation and its arguments.

pub fn get_distribution_latest_cache_reset(
&self
) -> GetDistributionLatestCacheReset<C, M, R>

Constructs a fluent builder for the GetDistributionLatestCacheReset operation.

See GetDistributionLatestCacheReset for more information about the operation and its arguments.

pub fn get_distribution_metric_data(&self) -> GetDistributionMetricData<C, M, R>

Constructs a fluent builder for the GetDistributionMetricData operation.

See GetDistributionMetricData for more information about the operation and its arguments.

pub fn get_distributions(&self) -> GetDistributions<C, M, R>

Constructs a fluent builder for the GetDistributions operation.

See GetDistributions for more information about the operation and its arguments.

pub fn get_domain(&self) -> GetDomain<C, M, R>

Constructs a fluent builder for the GetDomain operation.

See GetDomain for more information about the operation and its arguments.

pub fn get_domains(&self) -> GetDomains<C, M, R>

Constructs a fluent builder for the GetDomains operation.

See GetDomains for more information about the operation and its arguments.

pub fn get_export_snapshot_records(&self) -> GetExportSnapshotRecords<C, M, R>

Constructs a fluent builder for the GetExportSnapshotRecords operation.

See GetExportSnapshotRecords for more information about the operation and its arguments.

pub fn get_instance(&self) -> GetInstance<C, M, R>

Constructs a fluent builder for the GetInstance operation.

See GetInstance for more information about the operation and its arguments.

pub fn get_instance_access_details(&self) -> GetInstanceAccessDetails<C, M, R>

Constructs a fluent builder for the GetInstanceAccessDetails operation.

See GetInstanceAccessDetails for more information about the operation and its arguments.

pub fn get_instance_metric_data(&self) -> GetInstanceMetricData<C, M, R>

Constructs a fluent builder for the GetInstanceMetricData operation.

See GetInstanceMetricData for more information about the operation and its arguments.

pub fn get_instance_port_states(&self) -> GetInstancePortStates<C, M, R>

Constructs a fluent builder for the GetInstancePortStates operation.

See GetInstancePortStates for more information about the operation and its arguments.

pub fn get_instances(&self) -> GetInstances<C, M, R>

Constructs a fluent builder for the GetInstances operation.

See GetInstances for more information about the operation and its arguments.

pub fn get_instance_snapshot(&self) -> GetInstanceSnapshot<C, M, R>

Constructs a fluent builder for the GetInstanceSnapshot operation.

See GetInstanceSnapshot for more information about the operation and its arguments.

pub fn get_instance_snapshots(&self) -> GetInstanceSnapshots<C, M, R>

Constructs a fluent builder for the GetInstanceSnapshots operation.

See GetInstanceSnapshots for more information about the operation and its arguments.

pub fn get_instance_state(&self) -> GetInstanceState<C, M, R>

Constructs a fluent builder for the GetInstanceState operation.

See GetInstanceState for more information about the operation and its arguments.

pub fn get_key_pair(&self) -> GetKeyPair<C, M, R>

Constructs a fluent builder for the GetKeyPair operation.

See GetKeyPair for more information about the operation and its arguments.

pub fn get_key_pairs(&self) -> GetKeyPairs<C, M, R>

Constructs a fluent builder for the GetKeyPairs operation.

See GetKeyPairs for more information about the operation and its arguments.

pub fn get_load_balancer(&self) -> GetLoadBalancer<C, M, R>

Constructs a fluent builder for the GetLoadBalancer operation.

See GetLoadBalancer for more information about the operation and its arguments.

pub fn get_load_balancer_metric_data(
&self
) -> GetLoadBalancerMetricData<C, M, R>

Constructs a fluent builder for the GetLoadBalancerMetricData operation.

See GetLoadBalancerMetricData for more information about the operation and its arguments.

pub fn get_load_balancers(&self) -> GetLoadBalancers<C, M, R>

Constructs a fluent builder for the GetLoadBalancers operation.

See GetLoadBalancers for more information about the operation and its arguments.

pub fn get_load_balancer_tls_certificates(
&self
) -> GetLoadBalancerTlsCertificates<C, M, R>

Constructs a fluent builder for the GetLoadBalancerTlsCertificates operation.

See GetLoadBalancerTlsCertificates for more information about the operation and its arguments.

pub fn get_operation(&self) -> GetOperation<C, M, R>

Constructs a fluent builder for the GetOperation operation.

See GetOperation for more information about the operation and its arguments.

pub fn get_operations(&self) -> GetOperations<C, M, R>

Constructs a fluent builder for the GetOperations operation.

See GetOperations for more information about the operation and its arguments.

pub fn get_operations_for_resource(&self) -> GetOperationsForResource<C, M, R>

Constructs a fluent builder for the GetOperationsForResource operation.

See GetOperationsForResource for more information about the operation and its arguments.

pub fn get_regions(&self) -> GetRegions<C, M, R>

Constructs a fluent builder for the GetRegions operation.

See GetRegions for more information about the operation and its arguments.

pub fn get_relational_database(&self) -> GetRelationalDatabase<C, M, R>

Constructs a fluent builder for the GetRelationalDatabase operation.

See GetRelationalDatabase for more information about the operation and its arguments.

pub fn get_relational_database_blueprints(
&self
) -> GetRelationalDatabaseBlueprints<C, M, R>

Constructs a fluent builder for the GetRelationalDatabaseBlueprints operation.

See GetRelationalDatabaseBlueprints for more information about the operation and its arguments.

pub fn get_relational_database_bundles(
&self
) -> GetRelationalDatabaseBundles<C, M, R>

Constructs a fluent builder for the GetRelationalDatabaseBundles operation.

See GetRelationalDatabaseBundles for more information about the operation and its arguments.

pub fn get_relational_database_events(
&self
) -> GetRelationalDatabaseEvents<C, M, R>

Constructs a fluent builder for the GetRelationalDatabaseEvents operation.

See GetRelationalDatabaseEvents for more information about the operation and its arguments.

pub fn get_relational_database_log_events(
&self
) -> GetRelationalDatabaseLogEvents<C, M, R>

Constructs a fluent builder for the GetRelationalDatabaseLogEvents operation.

See GetRelationalDatabaseLogEvents for more information about the operation and its arguments.

pub fn get_relational_database_log_streams(
&self
) -> GetRelationalDatabaseLogStreams<C, M, R>

Constructs a fluent builder for the GetRelationalDatabaseLogStreams operation.

See GetRelationalDatabaseLogStreams for more information about the operation and its arguments.

pub fn get_relational_database_master_user_password(
&self
) -> GetRelationalDatabaseMasterUserPassword<C, M, R>

Constructs a fluent builder for the GetRelationalDatabaseMasterUserPassword operation.

See GetRelationalDatabaseMasterUserPassword for more information about the operation and its arguments.

pub fn get_relational_database_metric_data(
&self
) -> GetRelationalDatabaseMetricData<C, M, R>

Constructs a fluent builder for the GetRelationalDatabaseMetricData operation.

See GetRelationalDatabaseMetricData for more information about the operation and its arguments.

pub fn get_relational_database_parameters(
&self
) -> GetRelationalDatabaseParameters<C, M, R>

Constructs a fluent builder for the GetRelationalDatabaseParameters operation.

See GetRelationalDatabaseParameters for more information about the operation and its arguments.

pub fn get_relational_databases(&self) -> GetRelationalDatabases<C, M, R>

Constructs a fluent builder for the GetRelationalDatabases operation.

See GetRelationalDatabases for more information about the operation and its arguments.

pub fn get_relational_database_snapshot(
&self
) -> GetRelationalDatabaseSnapshot<C, M, R>

Constructs a fluent builder for the GetRelationalDatabaseSnapshot operation.

See GetRelationalDatabaseSnapshot for more information about the operation and its arguments.

pub fn get_relational_database_snapshots(
&self
) -> GetRelationalDatabaseSnapshots<C, M, R>

Constructs a fluent builder for the GetRelationalDatabaseSnapshots operation.

See GetRelationalDatabaseSnapshots for more information about the operation and its arguments.

pub fn get_static_ip(&self) -> GetStaticIp<C, M, R>

Constructs a fluent builder for the GetStaticIp operation.

See GetStaticIp for more information about the operation and its arguments.

pub fn get_static_ips(&self) -> GetStaticIps<C, M, R>

Constructs a fluent builder for the GetStaticIps operation.

See GetStaticIps for more information about the operation and its arguments.

pub fn import_key_pair(&self) -> ImportKeyPair<C, M, R>

Constructs a fluent builder for the ImportKeyPair operation.

See ImportKeyPair for more information about the operation and its arguments.

pub fn is_vpc_peered(&self) -> IsVpcPeered<C, M, R>

Constructs a fluent builder for the IsVpcPeered operation.

See IsVpcPeered for more information about the operation and its arguments.

pub fn open_instance_public_ports(&self) -> OpenInstancePublicPorts<C, M, R>

Constructs a fluent builder for the OpenInstancePublicPorts operation.

See OpenInstancePublicPorts for more information about the operation and its arguments.

pub fn peer_vpc(&self) -> PeerVpc<C, M, R>

Constructs a fluent builder for the PeerVpc operation.

See PeerVpc for more information about the operation and its arguments.

pub fn put_alarm(&self) -> PutAlarm<C, M, R>

Constructs a fluent builder for the PutAlarm operation.

See PutAlarm for more information about the operation and its arguments.

pub fn put_instance_public_ports(&self) -> PutInstancePublicPorts<C, M, R>

Constructs a fluent builder for the PutInstancePublicPorts operation.

See PutInstancePublicPorts for more information about the operation and its arguments.

pub fn reboot_instance(&self) -> RebootInstance<C, M, R>

Constructs a fluent builder for the RebootInstance operation.

See RebootInstance for more information about the operation and its arguments.

pub fn reboot_relational_database(&self) -> RebootRelationalDatabase<C, M, R>

Constructs a fluent builder for the RebootRelationalDatabase operation.

See RebootRelationalDatabase for more information about the operation and its arguments.

pub fn register_container_image(&self) -> RegisterContainerImage<C, M, R>

Constructs a fluent builder for the RegisterContainerImage operation.

See RegisterContainerImage for more information about the operation and its arguments.

pub fn release_static_ip(&self) -> ReleaseStaticIp<C, M, R>

Constructs a fluent builder for the ReleaseStaticIp operation.

See ReleaseStaticIp for more information about the operation and its arguments.

pub fn reset_distribution_cache(&self) -> ResetDistributionCache<C, M, R>

Constructs a fluent builder for the ResetDistributionCache operation.

See ResetDistributionCache for more information about the operation and its arguments.

pub fn send_contact_method_verification(
&self
) -> SendContactMethodVerification<C, M, R>

Constructs a fluent builder for the SendContactMethodVerification operation.

See SendContactMethodVerification for more information about the operation and its arguments.

pub fn set_ip_address_type(&self) -> SetIpAddressType<C, M, R>

Constructs a fluent builder for the SetIpAddressType operation.

See SetIpAddressType for more information about the operation and its arguments.

pub fn set_resource_access_for_bucket(
&self
) -> SetResourceAccessForBucket<C, M, R>

Constructs a fluent builder for the SetResourceAccessForBucket operation.

See SetResourceAccessForBucket for more information about the operation and its arguments.

pub fn start_instance(&self) -> StartInstance<C, M, R>

Constructs a fluent builder for the StartInstance operation.

See StartInstance for more information about the operation and its arguments.

pub fn start_relational_database(&self) -> StartRelationalDatabase<C, M, R>

Constructs a fluent builder for the StartRelationalDatabase operation.

See StartRelationalDatabase for more information about the operation and its arguments.

pub fn stop_instance(&self) -> StopInstance<C, M, R>

Constructs a fluent builder for the StopInstance operation.

See StopInstance for more information about the operation and its arguments.

pub fn stop_relational_database(&self) -> StopRelationalDatabase<C, M, R>

Constructs a fluent builder for the StopRelationalDatabase operation.

See StopRelationalDatabase for more information about the operation and its arguments.

pub fn tag_resource(&self) -> TagResource<C, M, R>

Constructs a fluent builder for the TagResource operation.

See TagResource for more information about the operation and its arguments.

pub fn test_alarm(&self) -> TestAlarm<C, M, R>

Constructs a fluent builder for the TestAlarm operation.

See TestAlarm for more information about the operation and its arguments.

pub fn unpeer_vpc(&self) -> UnpeerVpc<C, M, R>

Constructs a fluent builder for the UnpeerVpc operation.

See UnpeerVpc for more information about the operation and its arguments.

pub fn untag_resource(&self) -> UntagResource<C, M, R>

Constructs a fluent builder for the UntagResource operation.

See UntagResource for more information about the operation and its arguments.

pub fn update_bucket(&self) -> UpdateBucket<C, M, R>

Constructs a fluent builder for the UpdateBucket operation.

See UpdateBucket for more information about the operation and its arguments.

pub fn update_bucket_bundle(&self) -> UpdateBucketBundle<C, M, R>

Constructs a fluent builder for the UpdateBucketBundle operation.

See UpdateBucketBundle for more information about the operation and its arguments.

pub fn update_container_service(&self) -> UpdateContainerService<C, M, R>

Constructs a fluent builder for the UpdateContainerService operation.

See UpdateContainerService for more information about the operation and its arguments.

pub fn update_distribution(&self) -> UpdateDistribution<C, M, R>

Constructs a fluent builder for the UpdateDistribution operation.

See UpdateDistribution for more information about the operation and its arguments.

pub fn update_distribution_bundle(&self) -> UpdateDistributionBundle<C, M, R>

Constructs a fluent builder for the UpdateDistributionBundle operation.

See UpdateDistributionBundle for more information about the operation and its arguments.

pub fn update_domain_entry(&self) -> UpdateDomainEntry<C, M, R>

Constructs a fluent builder for the UpdateDomainEntry operation.

See UpdateDomainEntry for more information about the operation and its arguments.

pub fn update_load_balancer_attribute(
&self
) -> UpdateLoadBalancerAttribute<C, M, R>

Constructs a fluent builder for the UpdateLoadBalancerAttribute operation.

See UpdateLoadBalancerAttribute for more information about the operation and its arguments.

pub fn update_relational_database(&self) -> UpdateRelationalDatabase<C, M, R>

Constructs a fluent builder for the UpdateRelationalDatabase operation.

See UpdateRelationalDatabase for more information about the operation and its arguments.

pub fn update_relational_database_parameters(
&self
) -> UpdateRelationalDatabaseParameters<C, M, R>

Constructs a fluent builder for the UpdateRelationalDatabaseParameters operation.

See UpdateRelationalDatabaseParameters for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.